🚚rename: [자소서] 폴더 구조 변경 최종
- 폴더 구조를 2차적으로 변경하였습니다.

KAN-106
