+++ src/app/(route)/(personal-statement)/personal-statement/[id]/editing/_components/PSEditingBox.tsx
